Implement more user-based features

+++ src/components/pages/CounselorsPage.tsx
@@ -1,11 +1,17 @@
 // Copyright 2022 Social Fabric, LLC
 
-import React, { useContext, useState } from 'react';
+import React, { useContext, useEffect, useState } from 'react';
+import {
+  createPermission,
+  deletePermission,
+  updatePermission,
+} from '../../auth/permissions';
 import {
   Counselor,
   CounselorsContext,
   emptyCounselor,
 } from '../../data/counselors';
+import { LoggedInUserContext } from '../../data/users';
 import CreateCounselorModal from '../modals/CreateCounselorModal';
 import EditCounselorModal from '../modals/EditCounselorModal';
 import Navbar from '../navbar/Navbar';
@@ -16,32 +22,57 @@ const CounselorsPage: React.FC = () => {
     add: addCounselor,
     delete: deleteCounselor,
     update: updateCounselor,
-    // update: updateCounselor,
   } = useContext(CounselorsContext);
+  const { loggedInUser } = useContext(LoggedInUserContext);
+
   const [isCreateCounselorModalOpen, setIsCreateCounselorModalOpen] =
     useState<boolean>(false);
   const [isEditCounselorModalOpen, setIsEditCounselorModalOpen] =
     useState<boolean>(false);
   const [modalCounselor, setModalCounselor] =
     useState<Counselor>(emptyCounselor);
+  const [isCreateCounselorAllowed, setIsCreateCounselorAllowed] =
+    useState<boolean>(false);
+  const [isDeleteCounselorAllowed, setIsDeleteCounselorAllowed] =
+    useState<boolean>(false);
+  const [isUpdateCounselorAllowed, setIsUpdateCounselorAllowed] =
+    useState<boolean>(false);
+
+  useEffect(() => {
+    setIsCreateCounselorAllowed(
+      createPermission(loggedInUser.role, 'counselor')
+    );
+    setIsDeleteCounselorAllowed(
+      deletePermission(loggedInUser.role, 'counselor')
+    );
+    setIsUpdateCounselorAllowed(
+      updatePermission(loggedInUser.role, 'counselor')
+    );
+  }, [loggedInUser.role]);
 
   const handleCounselorAdded = (counselor: Counselor) => {
-    addCounselor(counselor);
+    if (isCreateCounselorAllowed) {
+      addCounselor(counselor);
+    }
   };
 
   const handleCounselorEdited = (counselor: Counselor) => {
-    updateCounselor(counselor);
+    if (isUpdateCounselorAllowed) {
+      updateCounselor(counselor);
+    }
   };
 
   const onCounselorDeleteClicked = (counselorToDelete: Counselor) => {
-    if (window.confirm('Delete this counselor?')) {
+    if (isDeleteCounselorAllowed && window.confirm('Delete this counselor?')) {
       deleteCounselor(counselorToDelete);
     }
   };
 
   const onCounselorEditClicked = (counselorToEdit: Counselor) => {
-    setModalCounselor(counselorToEdit);
-    setIsEditCounselorModalOpen(true);
+    if (isUpdateCounselorAllowed) {
+      setModalCounselor(counselorToEdit);
+      setIsEditCounselorModalOpen(true);
+    }
   };
 
   const onCounselorEmailClicked = (counselorToEmail: Counselor) => {
@@ -63,23 +94,29 @@ const CounselorsPage: React.FC = () => {
       </nav>
       <h1>Counselors</h1>
       <>
-        <button
-          type="button"
-          onClick={() => setIsCreateCounselorModalOpen(true)}
-        >
-          Add Counselor
-        </button>
-        <CreateCounselorModal
-          isOpen={isCreateCounselorModalOpen}
-          onCounselorAdded={handleCounselorAdded}
-          onClose={() => setIsCreateCounselorModalOpen(false)}
-        />
-        <EditCounselorModal
-          isOpen={isEditCounselorModalOpen}
-          onCounselorEdited={handleCounselorEdited}
-          onClose={() => setIsEditCounselorModalOpen(false)}
-          initialCounselor={modalCounselor}
-        />
+        {isCreateCounselorAllowed && (
+          <>
+            <button
+              type="button"
+              onClick={() => setIsCreateCounselorModalOpen(true)}
+            >
+              Add Counselor
+            </button>
+            <CreateCounselorModal
+              isOpen={isCreateCounselorModalOpen}
+              onCounselorAdded={handleCounselorAdded}
+              onClose={() => setIsCreateCounselorModalOpen(false)}
+            />
+          </>
+        )}
+        {isUpdateCounselorAllowed && (
+          <EditCounselorModal
+            isOpen={isEditCounselorModalOpen}
+            onCounselorEdited={handleCounselorEdited}
+            onClose={() => setIsEditCounselorModalOpen(false)}
+            initialCounselor={modalCounselor}
+          />
+        )}
         <CounselorsTable
           onDeleteClicked={onCounselorDeleteClicked}
           onEditClicked={onCounselorEditClicked}
