+++ src/components/pages/SchoolsPage.tsx
@@ -1,7 +1,13 @@
 // Copyright 2022 Social Fabric, LLC
 
-import React, { useContext, useState } from 'react';
+import React, { useContext, useEffect, useState } from 'react';
+import {
+  createPermission,
+  deletePermission,
+  updatePermission,
+} from '../../auth/permissions';
 import { emptySchool, School, SchoolsContext } from '../../data/schools';
+import { LoggedInUserContext } from '../../data/users';
 import CreateSchoolModal from '../modals/CreateSchoolModal';
 import EditSchoolModal from '../modals/EditSchoolModal';
 import Navbar from '../navbar/Navbar';
@@ -13,22 +19,40 @@ const SchoolsPage = () => {
     delete: deleteSchool,
     update: updateSchool,
   } = useContext(SchoolsContext);
+  const { loggedInUser } = useContext(LoggedInUserContext);
+
   const [isCreateSchoolModalOpen, setIsCreateSchoolModalOpen] =
     useState<boolean>(false);
   const [isEditSchoolModalOpen, setIsEditSchoolModalOpen] =
     useState<boolean>(false);
   const [modalSchool, setModalSchool] = useState<School>(emptySchool);
+  const [isCreateSchoolAllowed, setIsCreateSchoolAllowed] =
+    useState<boolean>(false);
+  const [isDeleteSchoolAllowed, setIsDeleteSchoolAllowed] =
+    useState<boolean>(false);
+  const [isUpdateSchoolAllowed, setIsUpdateSchoolAllowed] =
+    useState<boolean>(false);
+
+  useEffect(() => {
+    setIsCreateSchoolAllowed(createPermission(loggedInUser.role, 'school'));
+    setIsDeleteSchoolAllowed(deletePermission(loggedInUser.role, 'school'));
+    setIsUpdateSchoolAllowed(updatePermission(loggedInUser.role, 'school'));
+  }, [loggedInUser.role]);
 
   const handleSchoolAdded = (schoolToAdd: School) => {
-    addSchool(schoolToAdd);
+    if (isCreateSchoolAllowed) {
+      addSchool(schoolToAdd);
+    }
   };
 
   const handleSchoolEdited = (schoolToAdd: School) => {
-    updateSchool(schoolToAdd);
+    if (isUpdateSchoolAllowed) {
+      updateSchool(schoolToAdd);
+    }
   };
 
   const onSchoolDeleteClicked = (schoolToDelete: School) => {
-    if (window.confirm('Delete this user?')) {
+    if (isDeleteSchoolAllowed && window.confirm('Delete this user?')) {
       deleteSchool(schoolToDelete);
     }
   };
@@ -49,20 +73,29 @@ const SchoolsPage = () => {
       </nav>
       <h1>Schools</h1>
       <>
-        <button type="button" onClick={() => setIsCreateSchoolModalOpen(true)}>
-          Add School
-        </button>
-        <CreateSchoolModal
-          isOpen={isCreateSchoolModalOpen}
-          onSchoolAdded={handleSchoolAdded}
-          onClose={() => setIsCreateSchoolModalOpen(false)}
-        />
-        <EditSchoolModal
-          isOpen={isEditSchoolModalOpen}
-          onSchoolEdited={handleSchoolEdited}
-          onClose={() => setIsEditSchoolModalOpen(false)}
-          initialSchool={modalSchool}
-        />
+        {isCreateSchoolAllowed && (
+          <>
+            <button
+              type="button"
+              onClick={() => setIsCreateSchoolModalOpen(true)}
+            >
+              Add School
+            </button>
+            <CreateSchoolModal
+              isOpen={isCreateSchoolModalOpen}
+              onSchoolAdded={handleSchoolAdded}
+              onClose={() => setIsCreateSchoolModalOpen(false)}
+            />
+          </>
+        )}
+        {isUpdateSchoolAllowed && (
+          <EditSchoolModal
+            isOpen={isEditSchoolModalOpen}
+            onSchoolEdited={handleSchoolEdited}
+            onClose={() => setIsEditSchoolModalOpen(false)}
+            initialSchool={modalSchool}
+          />
+        )}
         <SchoolsTable
           onDeleteClicked={onSchoolDeleteClicked}
           onEditClicked={onSchoolEditClicked}
